+++ familyhelper-ui-view/packages/component-util/src/util/file.ts
@@ -1,0 +1,9 @@
+// noinspection JSUnusedGlobalSymbols
+
+export function parseFileExtension(fileName: string): string | '?' {
+  const index: number = fileName.lastIndexOf('.')
+  if (index < 0) {
+    return '?'
+  }
+  return fileName.substring(index + 1)
+}