+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- `component-ckeditor` 子模功能新增。
+  - 新增 `src/util/file.ts` 工具类。
+
 - `component-ckeditor` 子模块优化。
   - 添加 `dev` 指令以及相关代码，提升模块开发的便捷性。
   - `ckeditor5` 相关代码使用 `^45.0.0` 版本进行重写。